Create compartment sub sessions (Recidiviz/recidiviz-data#17004)
## Description of the change

This PR refactors `compartment_sessions` to be built off of a new
intermediate view called `compartment_sub_sessions`. Due to query
complexity limitations there is also a 2nd intermediate view upstream of
`compartment_sub_sessions` called
`compartment_sub_sessions_preprocessed` that is created.

`compartment_sub_sessions` contains all of the inference logic that was
previously in `compartment_sessions`. It is a view that has full date
coverage for each person with a new record that is created any time any
attribute changes. This includes the attributes "age" and
"assessment_score" which are not part of `dataflow_sessions`.

This new intermediate view will allows us to deprecate
`compartment_sessions_unnested` (for population metrics) and will also
give us more granular metrics as part of unified spans.

The output in `compartment_sessions` is identical with the exception of
the `age_end` field, which is now no longer hydrated for active inferred
compartments (people that are currently in a compartment not represented
in our population data). The current age of someone at liberty is not an
attribute that is really needed, and including it in within this new
framework would require some additional logic.
[This](https://docs.google.com/spreadsheets/d/1b0bR9VDmu8NREN4jydTrtFqfa7IS-EOOexBPXchuPuo/edit?usp=sharing)
sheet demonstrates the `compartment_sessions` match rate between the
current version and the newly refactored version.

And the following screen shot shows the materialization time for
updating all views upstream of `compartment_sessions`. The two new
sub-sessions views together take about 2.5 minutes to materialize.

<img width="807" alt="Screen Shot 2022-12-08 at 1 43 30 PM"
src="https://user-images.githubusercontent.com/6909506/206575944-1c5f13b7-98af-4ecf-9940-5b02efbedc2c.png">

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#16628

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cd901bc6a1cb5b02b42565f3f68878eca4526351