[US_IX] Revise state charge to link correctly to sentences (Recidiviz/recidiviz-data#17403)
## Description of the change

Revise state charge to link correctly to sentences, sepcifically:
- Pull `SentenceOrderCategoryId` into the ingest view and revise the
YAML mapping to attach to the appropriate type of sentence
(incarceration vs supervision) based on `SentenceOrderCategoryId`.
- Pull in the `SentenceId` for the _sentence_ attached to the same
sentence order as each charge to construct a sentence `external_id` we
can use to attach a charge to a sentence.
- For context on the original issue, sentences and charges come from the
same table `scl_Sentence` and are separate rows in that table (we
differentiate which rows are charges vs sentences using
`SentenceLinkClassId`) but to link charges to sentences we were
accidentally using the `SentenceId` for the charge row rather than the
`SentenceId` for the sentence row
- Updates all associated tests for the above changes

This should address both issues of not having charge <> supervision
sentence associations as well as not seeing charges attached to open
incarceration sentences.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e7bbea26c2e6c9c7b2ff7bfc0f665ea019e1a816