[SFTP] Adds steps for each state to check config and also set ingest lock (Recidiviz/recidiviz-data#17331)
## Description of the change

This PR sets up 3 things:
1. Allows for a specification of a state to have certain environments
supported for SFTP (to allow for Michigan to only be supported on
production).
2. Check that the state exists in a GCS config that houses which states
have Airflow SFTP enabled (which we'll remove eventually once we switch
everyone over).
3. Sets and releases various locks in the bucket to start ingest for
that state in the primary instance (for now).

As part of this, I had to move `GCSFileSystemImpl` to its own file to
minimize the dependency hell that we would run into with it. I tested
this all in the experimental DAG and we see the following results:

<img width="1407" alt="Screen Shot 2022-12-16 at 8 58 34 AM"
src="https://user-images.githubusercontent.com/5348488/208150346-0412b8a4-6057-4c71-a386-62ef17ab81b5.png">
<img width="1846" alt="Screen Shot 2022-12-16 at 8 59 36 AM"
src="https://user-images.githubusercontent.com/5348488/208150359-ca48ce8a-953c-48ef-a063-513fbe8b288c.png">

## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#17278

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: 2595f0d0c6f05392b3f57b8848259bbabbc406ea