[US_IX] Add mapping for bench warrant sessions (Recidiviz/recidiviz-data#17288)
Remap the supervision type / compartment level 2 in `dataflow_sessions`
for US_IX based on the supervising officer. It might make sense to have
this in ingest or calc instead of sessions (TODO Recidiviz/recidiviz-data#17392), but I've added
it to sessions as a quick fix to unblock some Workflows task eligibility
validation work. This update increases the sessions compartment level 2
match rate from 75% to 84% for supervision sessions open on 2022-11-08.

GitOrigin-RevId: b4699a1443a62065ccfccadaf48d560476d7afc9