[SFTP] Adding a FindSftpFilesToDownloadOperator and a custom SFTPHook (Recidiviz/recidiviz-data#17346)
## Description of the change

This adds a `FindSftpFilesToDownloadOperator` which makes a connection
to the SFTP server and determines the files that the server contains to
download. Note: we have the time limit currently and that will be
replaced once we're confident that the operations database can replace
this (which will be implemented later in a double-writing step as part
of the plan to switch to Airflow).

<img width="1450" alt="Screen Shot 2022-12-20 at 11 02 44 PM"
src="https://user-images.githubusercontent.com/5348488/209017831-4073a2e2-8088-47be-9239-3749ba0d17dc.png">

## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#12130
Closes Recidiviz/recidiviz-data#12134

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3b094d0363db10a00be97090f9cf6b3ecf94e060