[US_ND] Add INTERNAL_UNKNOWN supervision_type when admission reason is temp custody for commitments from supervision (Recidiviz/recidiviz-data#17507)
## Description of the change

This PR adds a new condition to
`get_pre_incarceration_supervision_type_from_ip_admission_reason` for
the US_ND commitment from supervision delegate that will return
INTERNAL_UNKNOWN when the admission reason is TEMPORARY_CUSTODY. Without
this condition we were seeing the dataflow job fail on an unhandled enum
error.

This is related to the change in
https://github.com/Recidiviz/recidiviz-data/issues/17370

Sandbox pipeline is running now!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#17487

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 079070efbb1b66807cd2e0537b04674297204da7