[US_OR] Adding field descriptions (Recidiviz/recidiviz-data#17509)
## Description of the change

While creating the Incarceration Periods I saw a few more fields that
might be pulled in. Adding the descriptions here so they can be used in
local views.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ca1105d95c548a22458a31f890893e73b3ac5592