[Justice Counts] Law enforcement funding metric (Recidiviz/recidiviz-data#17490)
## Description of the change

This metric deprecates the law enforcement funding metric and replaces
it with the new funding metric. This PR is rather large because so many
of our tests reference the law enforcement annual budget metric. I tried
my best to add annotations to the PR to explain the changes.

## Related issues

Related to Recidiviz/recidiviz-data#17241

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 98bb5bf2ea3fa08870018d5fa4369168575a72e7