[US_TN] Selecting first CountNumber for ISC Consecutive sentences associated with multiple counts in same case to unblock rerun (Recidiviz/recidiviz-data#17531)
## Description of the change

TN rerun is blocked because there are a few cases where an ISC sentence
is consecutive to multiple Count numbers for the same sentence in
ISCRelatedSentence. To unblock the rerun for now, we add in a row_number
to select the first count of the sentence. However, we will also follow
up with TN to ask how to handle these situations best in the future.

This query used to see all duplicates that were removed.
```
SELECT DISTINCT *
FROM `recidiviz-staging.gheidkamp_tn_scratch.SCCC_v2_original`
EXCEPT DISTINCT
SELECT DISTINCT *
FROM `recidiviz-staging.gheidkamp_tn_scratch.SCCC_v2_isc_correct`
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#17480

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fbaf163b8a155ce4fcb452f8dd38cde3e3f7249c